// packages/backend/types/database.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
// ...
  public: {
    Tables: {
      presentation_events: {
        Row: {
          created_at: string
          created_by: string | null
          created_by_alias: string | null
          created_by_anon_uuid: string | null
          id: number
          is_public: boolean
          presentation: number
          type: Database["public"]["Enums"]["event_type"]
          value: Json | null
        }
        Insert: {
          created_at?: string
          created_by?: string | null
          created_by_alias?: string | null
          created_by_anon_uuid?: string | null
          id?: number
          is_public?: boolean
          presentation: number
          type: Database["public"]["Enums"]["event_type"]
          value?: Json | null
        }
        Update: {
          created_at?: string
          created_by?: string | null
          created_by_alias?: string | null
          created_by_anon_uuid?: string | null
          id?: number
          is_public?: boolean
          presentation?: number
          type?: Database["public"]["Enums"]["event_type"]
          value?: Json | null
        }
        Relationships: [
          {
            foreignKeyName: "presentation_events_created_by_fkey"
            columns: ["created_by"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "presentation_events_presentation_fkey"
            columns: ["presentation"]
            isOneToOne: false
            referencedRelation: "presentations"
            referencedColumns: ["id"]
          },
        ]
      }
      presentations: {
        Row: {
          created_at: string
          description: string | null
          id: number
          join_code: string | null
          lc_status: Database["public"]["Enums"]["presentation_lifecycle_status"]
          presenter: string
          title: string | null
        }
        Insert: {
          created_at?: string
          description?: string | null
          id?: number
          join_code?: string | null
          lc_status?: Database["public"]["Enums"]["presentation_lifecycle_status"]
          presenter?: string
          title?: string | null
        }
        Update: {
          created_at?: string
          description?: string | null
          id?: number
          join_code?: string | null
          lc_status?: Database["public"]["Enums"]["presentation_lifecycle_status"]
          presenter?: string
          title?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "presentation_presenter_fkey"
            columns: ["presenter"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
      profiles: {
        Row: {
          avatar_url: string | null
          full_name: string | null
          id: string
          updated_at: string | null
          username: string | null
          website: string | null
        }
        Insert: {
          avatar_url?: string | null
          full_name?: string | null
          id: string
          updated_at?: string | null
          username?: string | null
          website?: string | null
        }
        Update: {
          avatar_url?: string | null
          full_name?: string | null
          id?: string
          updated_at?: string | null
          username?: string | null
          website?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "profiles_id_fkey"
            columns: ["id"]
            isOneToOne: true
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
    }
// ...
    CompositeTypes: {
      generic_acknowledgement_type: {
        entity: string | null
        id: number | null
        acknowledged_at: string | null
      }
      presentation_peek_type: {
        id: number | null
        title: string | null
        description: string | null
        lc_status:
          | Database["public"]["Enums"]["presentation_lifecycle_status"]
          | null
        presenter_username: string | null
        presenter_fullname: string | null
      }
    }
  }
// ...
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never